[feat/LIG-10] 1. sliding and session window done; 2. optimize imports; 3. add some uts
