+++ src/worker/src/main.rs
@@ -1,6 +1,6 @@
 use common::utils::{self, get_env};
 use proto::worker::task_worker_api_server::TaskWorkerApiServer;
-use std::{fs, time::Duration};
+use std::fs;
 use stream::initialize_v8;
 use tonic::transport::Server;
